flask_back_end/requirements.txt
Flask-PyMongo
PyJWT
flask_uploads
flask-login
requests
